requirements.txt
asgiref==3.8.1
Django==5.2
image==1.5.33
joblib==1.4.2
numpy==2.2.4
opencv-python==4.11.0.86
pillow==11.1.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
